requirements.txt
ortools.sat.python
pandas
datetime
tabulate
json
flask
fpdf
pytest